refactoring base structure finished
* removed get points from region (unused)
* separate utils into time based and io based functions
* turn plot and mask into functions for both lines and regions2d
