+++ echoregions/__init__.py
@@ -1,5 +1,6 @@
 from .utils.api import read_evl, read_evr
-from .utils.utils import parse_simrad_fname_time, parse_time
+from .utils.time import parse_simrad_fname_time, parse_time
+from .utils.io import from_JSON, check_file, validate_path
 from .lines.lines import Lines
 from .regions2d.regions2d import Regions2D
 
@@ -8,6 +9,9 @@ __all__ = [
     "read_evr",
     "parse_simrad_fname_time",
     "parse_time",
+    "from_JSON",
+    "check_file",
+    "validate_path",
     "Lines",
     "Regions2D",
 ]
