+++ echoregions/lines/__init__.py
@@ -1,5 +1,5 @@
 from .lines import Lines
-from .lines_plot import LinesPlotter
+from .lines_mask import lines_mask
 from .lines_parser import parse_line_file
 
-__all__ = ["Lines", "LinesPlotter", "parse_line_file"]
+__all__ = ["Lines", "lines_mask", "parse_line_file"]
