+++ echoregions/lines/lines_mask.py
@@ -1,0 +1,3 @@
+def lines_mask():
+    # TODO Implment Lines Mask
+    return
