+++ echoregions/regions2d/__init__.py
@@ -1,6 +1,5 @@
 from .regions2d import Regions2D
-from .regions2d_mask import Regions2DMasker
-from .regions2d_plot import Regions2DPlotter
+from .regions2d_mask import regions2d_mask
 from .regions2d_parser import parse_regions_file
 
-__all__ = ["Lines", "LinesPlotter", "parse_line_file"]
+__all__ = ["Regions2D", "regions2d_mask", "parse_regions_file"]
