+++ echoregions/utils/__init__.py
@@ -1,3 +1,4 @@
-from .utils import from_JSON, parse_simrad_fname_time, parse_time
+from .io import from_JSON, check_file
+from .time import parse_simrad_fname_time, parse_time
 
-__all__ = ["from_JSON", "parse_simrad_fname_time", "parse_time"]
+__all__ = ["from_JSON", "check_file", "parse_simrad_fname_time", "parse_time"]
